ml_service/requirements.txt
openai-whisper
phonemizer
python-Levenshtein
flask
torch
